добавлен прогресс бар для воспроизводящихся подкастов

+++ app/src/main/java/ru/anydevprojects/simplepodcastapp/core/mediaPlayerControl/domain/MediaData.kt
@@ -10,6 +10,6 @@ sealed interface MediaData {
         val id: Long,
         val title: String,
         val imageUrl: String,
-        val totalDurationMs: Int
+        val totalDurationMs: Long
     ) : MediaData
 }
